* Added struct layout attributes * added or corrected type identifier for some enums * changed Unknown15 back to byte array * added pack sizes for some structs * Bugfix SramSizes: corrected wrong value of Unknown17B * SramFileSoE: moved SizeCkeds to OnLoading callback method

+++ Models/Enums/LandingLocation.cs
@@ -1,6 +1,6 @@
 namespace RosettaStone.Sram.SoE.Models.Enums
 {
-	public enum LandingLocation : ushort
+	public enum LandingLocation : byte
 	{
 		Unknown1 = 0,
 		Unknown2 = 2,
